Update tikz template, fix some bugs, add graph README
The tikz template for graphs now uses layered layout by default,
and also includes all of the tikz libraries and a document class,
so it's a standalone document.

I had some trouble running tests, because there were classes that
were in the wrong place, so I moved those. I also added a `pass`
where it was needed, because otherwise the indentation was messed
up.

I added a graph readme, GRAPH_VIS.md, and also some example images
of the graphs that are produced.