Reformat + add/update documentation
Refactored some code to make it more consistent

Added convenience functions for outputting TeX and also dot output
files.

Added documentation to the tikz and dot graph functions in graph.py.

Renamed graph_vis_tikz_template.py to graph_vis_support, since I
added some additional support for dot graph output. It turns out that
there are like 20+ output formats for dot, and some of them are obsolete,
but if you don't specify an extension, some operating systems won't know
what kind of file it's dealing with. So now it automatically adds an
extension if it's not there, in the cases where a widely accepted extension
exists. This behavior can be turned off by providing a different value for
a optional parameter.