Make some headway on revisions of GUI
I've gotten it to ask you questions, and it even fills things out
based on the questions. However, there are some gray areas, like
what if for the answer to a question, there is only one reasonable
family, but you have a mixed effects model and that family is not
supported?